requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pydantic
sqlalchemy
psycopg2
python-dotenv==1.0.0
redis